# openapi.yaml
openapi: 3.1.0
info:
  title: Bookstore API
  version: 1.0.0
  description: Basic demo e-commerce API
...
paths:
  /books:
    description: Returns all books
    get:
      summary: Gets all books data
      description: Returns all books in the store
      responses:
        "200":
          description: An array of book objects
          content:
            application/json:
              examples:
                books:
                  value: |-
                    [
                        {
                            "id": 1,
                            "title": "The Return of the King",
                            "author": "J.R.R Tolkien",
                            "num_of_pages": 1567,
                            "genre": "Fantasy",
                            "price": "£6.99"
                        },
                        {
                            "id": 2,
                            "title": "Israel-Palestine",
                            "author": "Alain Gresh",
                            "num_of_pages": 300,
                            "genre": "Non-fiction",
                            "price": "£8.00"
                        }
                    ]
        "400":
          description: Message containing the error stack
      tags:
        - books
    post:
      summary: Creates a new book (admin only)
      description: Creates a new book to add into the store
      requestBody:
        description: >-
          A JSON object containing the title, author, num_of_pages, genre and
          price of a book to be added
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        "201":
          description: The book has been successfully added to the store
          content:
            text/plain; charset=utf-8:
              examples:
                Successful book creation:
                  value: Book added!
        "400":
          description: Message containing the error stack
      tags:
        - books
        - admin
  /books/{bookId}:
    description: Returns a book by its given id
    get:
      summary: Returns a book with matching ID
      parameters:
        - in: path
          name: bookId
          schema:
            type: integer
          required: true
          description: Numeric ID of the book to get
      description: Returns a book by its given id
      responses:
        "200":
          description: Returns one book object
          content:
            application/json:
              examples:
                /books/1:
                  value: |-
                    [
                        {
                            "id": 1,
                            "title": "The Return of the King",
                            "author": "J.R.R Tolkien",
                            "num_of_pages": 1567,
                            "genre": "Fantasy",
                            "price": "£6.99"
                        }
                    ]
        "400":
          description: Message containing the error stack
        "404":
          description: No book exists matching that ID
          content:
            text/plain; charset=utf-8:
              examples:
                "Book id not matching ":
                  value: Book not found
        default:
          description: Default error sample response
      tags:
        - books
    put:
      summary: Update the price of a book (admin only)
      parameters:
        - in: path
          name: bookId
          schema:
            type: integer
          required: true
          description: Numeric ID of the book to put
      description: This operation updates only the price of a book in the store
      requestBody:
        description: A JSON object containing the to-be-updated price of the book
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                price:
                  type: number
                  example: 8.99
              required:
                - price
      responses:
        "200":
          description: Successful update of the price of a book
          content:
            text/plain; charset=utf-8:
              examples:
                Successful update:
                  value: "Book with id 1: price updated"
        "400":
          description: Message containing the error stack
        "404":
          description: No book in the database matched the given ID
          content:
            text/plain; charset=utf-8:
              examples:
                No book found with requested ID:
                  value: Book not found
      tags:
        - books
        - admin
    delete:
      summary: Deletes a book with matching ID (admin only)
      parameters:
        - in: path
          name: bookId
          schema:
            type: integer
          required: true
          description: Numeric ID of the book to delete
      description: >-
        Deletes a book from the database permanently whose ID matches the
        endpoint book ID
      responses:
        "204":
          description: Successful deletion
          content:
            text/plain; charset=utf-8:
              examples:
                Successful deletion:
                  value: Book deleted!
        default:
          description: Default error sample response
      tags:
        - books
        - admin
  /users/all:
    description: Returns all users
    get:
      summary: Gets all users data (admin only)
      description: Returns an array of all the users on the system
      responses:
        "200":
          description: Successful retrieval of users data
          content:
            application/json:
              examples:
                /users:
                  value: |-
                    [
                        {
                            "id": 1,
                            "username": "Guest",
                            "password": "$2b$10$MVv2fdnUR2xsSBpgDgLGiuXvDgMjbpBa.4hPy/0rT9SUq1f60LdMq",
                            "email": "guest@gmail.com",
                            "isadmin": "true"
                        }
                    ]
        "400":
          description: Message containing the error stack
      tags:
        - users
        - admin
  /users:
    get:
      summary: Returns the user's data
      description: Returns the user object for logged in user
      responses:
        "200":
          description: Successful retrieval of user object
          content:
            application/json:
              examples:
                /users:
                  value: |-
                    [
                        {
                            "id": 1,
                            "username": "Guest",
                            "password": "$2b$10$MVv2fdnUR2xsSBpgDgLGiuXvDgMjbpBa.4hPy/0rT9SUq1f60LdMq",
                            "email": "guest@gmail.com",
                            "isadmin": "true"
                        }
                    ]
        "400":
          description: Message containing the error stack
      tags:
        - users
    put:
      summary: Updates username and password for loggedIn user
      description: Allows a user to update their username and their password
      requestBody:
        description: A JSON object containing the new username and password of the user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: billJohnson1
                password:
                  type: string
                  example: password1
              required:
                - username
                - password
      responses:
        "200":
          description: Successful update
          content:
            text/plain; charset=utf-8:
              examples:
                /user:
                  value: You've updated your username and password!
        "400":
          description: Message containing the error stack
      tags:
        - users
  /users/register:
    summary: Creates a user
    description: Register a user
    post:
      summary: Register a new user
      description: >-
        Register a new user with a username, password (to be encrypted), and
        email, isadmin is default false
      requestBody:
        description: >-
          A JSON object containing the username, password and email of a user to
          be added
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: Successful registration
          content:
            application/json:
              examples:
                users/register:
                  value: Your registration was successful!
        "400":
          description: Message containing the error stack
        "404":
          description: User with the given email already exists
          content:
            text/plain; charset=utf-8:
              examples:
                Email already exists:
                  value: User with that email already exists!
      tags:
        - users
...
  /users/carts:
    description: Returns all carts belonging to the loggedIn user
    get:
      summary: Gets all cart with all its items belonging to the loggedIn user
      description: >-
        Returns a cart with all its items, alongside their title, price, author
        and quantity
      responses:
        "200":
          description: Successful retrieval of a cart
          content:
            application/json:
              examples:
                Successful retrieval of a cart:
                  value: |-
                    [
                        {
                            "title": "Israel-Palestine",
                            "author": "Alain Gresh",
                            "price": "£8.00",
                            "quantity": 3
                        },
                        {
                            "title": "The Return of the King",
                            "author": "J.R.R Tolkien",
                            "price": "£6.99",
                            "quantity": 2
                        }
                    ]
        "400":
          description: Message containing the error stack
        "404":
          description: USer id doesn't match any user in the database
          content:
            text/plain; charset=utf-8:
              examples:
                User not found (invalid user ID):
                  value: User not found
      tags:
        - carts
    post:
      summary: Creates a cart for the loggedIn user
      description: Creates a cart for the loggedIn user
      responses:
        "201":
          description: Successful creation of a cart
          content:
            text/plain; charset=utf-8:
              examples:
                Successful creation of a cart:
                  value: Cart created
        "400":
          description: Message containing the error stack
      tags:
        - carts
...
components:
  schemas:
    Book:
      type: object
      properties:
        title:
          type: string
          maxLength: 40
          example: The Return of the King
        author:
          type: string
          maxLength: 50
          example: J.R.R Tolkien
        num_of_pages:
          type: integer
          example: 1567
        genre:
          type: string
          maxLength: 25
          example: Fantasy
        price:
          type: number
          example: 6.99
      required:
        - title
        - author
        - num_of_pages
        - genre
        - price
    User:
      type: object
      properties:
        username:
          type: string
          maxLength: 20
          example: Guest
        password:
          type: string
          example: password
        email:
          type: string
          example: guest@gmail.com
        isadmin:
          type: boolean
          example: true
      required:
        - username
        - password
        - email
        - isadmin